# openapi.yaml
openapi: 3.0.3
info:
  title: Gateway Zero API
  description: |
    Gateway Zero 是一个基于 Rust 和 Axum 框架开发的 AI API 网关系统。
    它提供统一的 OpenAI 兼容接口，支持多个 AI 提供商（OpenAI、Anthropic、智谱 AI 等），
    并具备负载均衡、令牌管理、用量统计等企业级功能。
  version: 0.1.0
  contact:
    name: Gateway Zero Team
servers:
  - url: http://localhost:8080
    description: 本地开发服务器
  - url: https://api.example.com
    description: 生产环境服务器
components:
  securitySchemes:
    # Client Token：外部调用 /v1/* 的 API Token（与 Admin Identity 严格区分）
    ClientToken:
      type: apiKey
      in: header
      name: Authorization
      description: "Client Token 认证（格式: Bearer <token>），用于访问 `/v1/*`"

    # Admin Identity：管理员身份（可来自 JWT AccessToken / TUI Session bearer / Web Session Cookie）
    AdminTuiSessionToken:
      type: http
      scheme: bearer
      description: 管理员身份（Admin Identity）的 TUI Session Token（bearer，但非 JWT）

    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: gw_session
      description: 管理员身份（Admin Identity）的 Web Session Cookie（HttpOnly）

    AccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "AccessToken（JWT），格式: Bearer <jwt>；用于访问 `/auth/me`、`/auth/change-password`、`/me/*` 以及 superadmin-only 的 `/admin/*`、`/providers/*`"
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: 错误代码 (如 http_error, not_found, rate_limited 等)
        message:
          type: string
          description: 错误详细信息

    # 请求日志
...
    ChatCompletionRequest:
      type: object
      properties:
        model:
          type: string
          description: "模型名称 (格式: provider/model 或 model)"
        messages:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [system, user, assistant]
              content:
                type: string
          description: 消息列表
        stream:
          type: boolean
          default: false
          description: 是否使用流式响应
        temperature:
          type: number
          format: float
          nullable: true
          description: 温度参数
        max_tokens:
          type: integer
          nullable: true
          description: 最大生成 tokens
        top_p:
          type: number
          format: float
          nullable: true
          description: Top-p 采样参数
      required:
        - model
        - messages

    # 聊天补全响应
    ChatCompletionResponse:
      type: object
      properties:
        id:
          type: string
          description: 响应ID
        object:
          type: string
          default: chat.completion
        created:
          type: integer
          format: int64
          description: 创建时间戳
        model:
          type: string
          description: 使用的模型
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              message:
                type: object
                properties:
                  role:
                    type: string
                  content:
                    type: string
              finish_reason:
                type: string
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer

    # 令牌余额响应（/v1/token/balance）
...
paths:
...
  /v1/chat/completions:
    post:
      summary: 聊天补全
      description: 创建聊天补全请求，支持流式和非流式响应
      operationId: createChatCompletion
      tags:
        - Chat
      security:
        - ClientToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
            text/event-stream:
              schema:
                type: string
                description: SSE 流式响应
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...